refactor(OpenGL): splitting the systems into their respective src files
