+++ src/plugin/opengl/src/system/BufferSystems.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Core.hpp"
+
+namespace ES::Plugin::OpenGL::System {
+void LoadGLMeshBuffer(ES::Engine::Core &core);
+void LoadGLSpriteBuffer(ES::Engine::Core &core);
+void LoadGLTextBuffer(ES::Engine::Core &core);
+} // namespace ES::Plugin::OpenGL::System
