+++ src/plugin/opengl/src/system/RenderSystems.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Core.hpp"
+
+namespace ES::Plugin::OpenGL::System {
+void RenderMeshes(ES::Engine::Core &core);
+void RenderText(ES::Engine::Core &core);
+void RenderSprites(ES::Engine::Core &core);
+} // namespace ES::Plugin::OpenGL::System
